添加查看卡组指令，还没实现

+++ lib/widget_msgbox.rb
@@ -69,7 +69,8 @@ class Widget_Msgbox < Window
     end
   end
   def clicked
-    @proc.call(@index) if @index
+    return if @index.nil?
+    @proc.call(@index) if @proc
     self.destroy
   end
 end
